add conftest and move device fixture there
