+++ tests/field/test_linear_actuator.py
@@ -12,25 +12,6 @@ from artist.util.configuration_classes import (
     KinematicDeviations,
 )
 
-
-@pytest.fixture(params=["cpu", "cuda"] if torch.cuda.is_available() else ["cpu"])
-def device(request: pytest.FixtureRequest) -> torch.device:
-    """
-    Return the device on which to initialize tensors.
-
-    Parameters
-    ----------
-    request : pytest.FixtureRequest
-        The pytest fixture used to consider different test cases.
-
-    Returns
-    -------
-    torch.device
-        The device on which to initialize tensors.
-    """
-    return torch.device(request.param)
-
-
 @pytest.fixture
 def deviation_parameters(device: torch.device) -> KinematicDeviations:
     """
