+++ tests/field/test_only_nurbs.py
@@ -4,24 +4,6 @@ import torch
 from artist.util.nurbs import NURBSSurface
 
 
-@pytest.fixture(params=["cpu", "cuda"] if torch.cuda.is_available() else ["cpu"])
-def device(request: pytest.FixtureRequest) -> torch.device:
-    """
-    Return the device on which to initialize tensors.
-
-    Parameters
-    ----------
-    request : pytest.FixtureRequest
-        The pytest fixture used to consider different test cases.
-
-    Returns
-    -------
-    torch.device
-        The device on which to initialize tensors.
-    """
-    return torch.device(request.param)
-
-
 def random_surface(
     e: torch.Tensor,
     n: torch.Tensor,
