+++ tests/field/test_rigid_body.py
@@ -11,25 +11,6 @@ from artist.util.configuration_classes import (
     KinematicOffsets,
 )
 
-
-@pytest.fixture(params=["cpu", "cuda"] if torch.cuda.is_available() else ["cpu"])
-def device(request: pytest.FixtureRequest) -> torch.device:
-    """
-    Return the device on which to initialize tensors.
-
-    Parameters
-    ----------
-    request : pytest.FixtureRequest
-        The pytest fixture used to consider different test cases.
-
-    Returns
-    -------
-    torch.device
-        The device on which to initialize tensors.
-    """
-    return torch.device(request.param)
-
-
 @pytest.fixture
 def actuator_configuration() -> ActuatorListConfig:
     """
