+++ tests/scene/test_light_source.py
@@ -7,24 +7,6 @@ from artist.scene import Sun
 from artist.util import config_dictionary
 
 
-@pytest.fixture(params=["cpu", "cuda"] if torch.cuda.is_available() else ["cpu"])
-def device(request: pytest.FixtureRequest) -> torch.device:
-    """
-    Return the device on which to initialize tensors.
-
-    Parameters
-    ----------
-    request : pytest.FixtureRequest
-        The pytest fixture used to consider different test cases.
-
-    Returns
-    -------
-    torch.device
-        The device on which to initialize tensors.
-    """
-    return torch.device(request.param)
-
-
 def calculate_expected(
     distribution_parameters: Dict[str, Any],
     further_parameters: Dict[str, int],
