+++ tests/util/test_wgs84_to_enu.py
@@ -4,24 +4,6 @@ import torch
 from artist.util import utils
 
 
-@pytest.fixture(params=["cpu", "cuda"] if torch.cuda.is_available() else ["cpu"])
-def device(request: pytest.FixtureRequest) -> torch.device:
-    """
-    Return the device on which to initialize tensors.
-
-    Parameters
-    ----------
-    request : pytest.FixtureRequest
-        The pytest fixture used to consider different test cases.
-
-    Returns
-    -------
-    torch.device
-        The device on which to initialize tensors.
-    """
-    return torch.device(request.param)
-
-
 @pytest.mark.parametrize(
     "wgs84_coordinates, reference_point, expected_enu_coordinates",
     [
